requirements_test.txt
pytest>=2.7.0
pytest-flask>=0.8.1
selenium>=2.45.0
